X4: additional mod name resolution enhancements
Initially we were assuming that the only cases where a mod's content id
would be different from the mod's folder name would be for mods which
have been migrated from Steam Workshop and had the "ws_xxxxxx" content
id. A member of the X4 community, kindly pointed out that this is not
the case, and other mods can also have this discreprancy. This commit
would ensure that:

1. If a mod already has the correct structure - use the archive
structure.
2. If a mod does not respect the correct structure (edge case) we would
attempt to read the correct value from the index files.
3. If all else fails - use the content id.