W3: temporary hack to avoid adding managed mods as external to LO page
We're currently investigating an issue causing the Load Order page to
wrongfully pick up managed mods as external - this seems to happen when
a user removes a mod which is part of a merge which was created by the
W3 script merger, and then proceeds to remove the merge through the
merger tool. (currently cannot reproduce)

This commit should prevent the above use case from happening.