W3: migrating merge inventories is no longer required
The script merger script is extracted into the existing directory of the
tool (if it's confirmed to already be installed). Given this fact, the
inventory migration would attempt to copy the file onto itself which
obviously will raise an error as seen in nexus-mods/vortex#6292
Removing it will ensure this does not happen.