W3fix:dummy installer to stop users installing script merger as a mod
It appears that users are trying to install the script merger tool as a
mod instead of a tool. The installation and configuration should've been
automatic starting from 1.2.12, the fact that users are trying to
install it as a mod may suggest that the download/installation had
failed. nexus-mods/vortex#6339