W3: stop blocking merges due to invalid mod xml
It appears that some mods may be distributed with invalid XML config
files. Although unfortunate and probably means that the mod won't work
as expected, we should probably not stop the entire deployment due to 1
faulty mod. We now report the issue to the user via a notification, and
just keep going.