KC:D fixed error on game activation if profile had become undefined
This error appears to be raised when a user switches to KC:D and the
extension attempts to update the load order for the currently active
profile which _should_ be available at this point, but for some reason
is undefined. Possibly the user had quickly switched between extensions?
(possibly removing the profile altogether?)

closes nexus-mods/vortex#7053