Small bugfixes
Removed QMM from the dashboard (it's no longer required) 
Added a check to the setup to ensure the QMods folder exists. (Prevents an error if you haven't run the game before modding it)
Check for the QModInstaller.dll file, not the QModManager.exe as the QMM team have said this may be removed entirely at some point.