MnB2: added protective code to cater for state weirdness
This is arguably a band aid fix for an issue that MUST be caused by some
sort of state corruption. The error manifests itself as a crash when we
attempt to populate our deployed mods cache (used for pretty much
everything in this extension) and the stagingFolder or mod installation
are undefined, although both use cases should be impossible given the
location where the crash occurs.

closes nexus-mods/vortex#7075