RE2/3/DMC5: games will now deactivate if failed re-wrapper registration
Previously it was possible for the re-wrapper to wrongly insert a game
into its "supported" container even if the QBMS extension was missing or
disabled. This modification is simply another sanity check to ensure
that user cannot mess around with their cache unless we confirm that
both the QBMS and RE-Wrapper extensions are fully functional.