W3: final few modifications to better support the priority manager
This basically adds the ability to switch between two different priority
assignment functionalities.

The prefix based assignment which is what we were doing up until now and
is the less restrictive assignment method as the user is able to assign
any priority value he wishes to his mod entries.

The position based assignment method which will always ensure that the
load order entries are assigned priorities based on their position in
the load order page in an incremental manner (e.g 1-40)