W3: Added merged data mismatch warning when uploading a collection
Merged scripts/data rely on their respective mods to be installed as
part of the collection. Unfortunately it's possible for the user to
remove a mod from a collection or highlighting it as "optional" and
forget to remove its corresponding merged script entry. This will
cause the collection to break on the user's end.

We added a check to ensure that all merged scripts/data uploaded as part
of a collection have their respective mods added to it and that they
are marked as "Required" - otherwise a dialog will inform the curator
of any missing/optional mods before he uploads the collection.