W3: fixed deployment errors for collections which include menu mod cache
The menu mod cache includes absolute paths pointing to a mod's
configuration (if applicable) which is used to apply/remove said config.

The absolute paths as exported by collections would function fine on the
curator's end, but not on the user's system - this is now fixed.
(even for collections that have already been published)

Additional fixes:
- Fixed duplicate load order entries for "locked" mods
- Fixed Disable/Enable all buttons not functioning correctly